splitted raycaster.c
