+++ Makefile
@@ -2,7 +2,9 @@
 PROJECT_NAME	= cub3d
 
 #--C FILES--# >
-C_FILES_MAIN	= cub3d.c error.c free_data.c raycaster.c draw.c	\
+C_FILES_MAIN	= cub3d.c error.c free_data.c raycaster.c data.c	\
+				  draw_minimap.c moves_1.c moves_2.c utils.c 		\
+				  raycasting_utils.c								\
 
 C_FILES_CHECK	= check_arguments.c get_map.c check_path.c			\
 				  check_arguments_1.c check_map.c check_map_utils.c	\
