--- asset/wall.xpm
@@ -1,43 +1,0 @@
-/* XPM */
-static char *_d4c83fd93874431ad4b37a661f0b494Axza5mZpOa8A7AXs[] = {
-/* columns rows colors chars-per-pixel */
-"32 32 5 1 ",
-"  c #001024",
-". c #021D43",
-"X c #0A3258",
-"o c #094C6A",
-"O c #357083",
-/* pixels */
-"                                ",
-"                                ",
-"                                ",
-"   oooooOOOXXooooooOOXXXooooo   ",
-"   oooooOOOXXooooooOOXXXooooo   ",
-"   XXoooooo..XXXooooo...XXXoo   ",
-"   XXoooooo..XXXooooo...XXXoo   ",
-"   XXoooooo..XXXooooo...XXXoo   ",
-"   ..XXXXXX........XX......XX   ",
-"   ..XXXXXX........XX......XX   ",
-"   ..XXXXXX........XX......XX   ",
-"   OOOOOOOOXXoooOOOOOXXXoooOO   ",
-"   OOOOOOOOXXoooOOOOOXXXoooOO   ",
-"   OOOOOoooXXooooooOOXXXooooo   ",
-"   OOOOOoooXXooooooOOXXXooooo   ",
-"   OOOOOoooXXooooooOOXXXooooo   ",
-"   oooooooo..XXXooooo...XXXoo   ",
-"   oooooooo..XXXooooo...XXXoo   ",
-"   oooooooo..XXXooooo...XXXoo   ",
-"   XXXXXXXX........XX......XX   ",
-"   XXXXXXXX........XX......XX   ",
-"   ooOOOOOOXXoooOOOOOXXXoooOO   ",
-"   ooOOOOOOXXoooOOOOOXXXoooOO   ",
-"   ooOOOOOOXXoooOOOOOXXXoooOO   ",
-"   oooooOOOXXooooooOOXXXooooo   ",
-"   oooooOOOXXooooooOOXXXooooo   ",
-"   oooooOOOXXooooooOOXXXooooo   ",
-"   XXoooooo..oooOOOOO...XXXoo   ",
-"   XXoooooo..oooOOOOO...XXXoo   ",
-"                                ",
-"                                ",
-"                                "
-};
