--- src/draw.c
@@ -1,107 +1,0 @@
-#include "../include/cub3d.h"
-
-/* int worldMap[mapWidth][mapHeight]=
-{
-  {1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1},
-  {1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1},
-  {1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1},
-  {1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1},
-  {1,0,0,0,0,0,2,2,2,2,2,0,0,0,0,3,0,3,0,3,0,0,0,1},
-  {1,0,0,0,0,0,2,0,0,0,2,0,0,0,0,0,0,0,0,0,0,0,0,1},
-  {1,0,0,0,0,0,2,0,0,0,2,0,0,0,0,3,0,0,0,3,0,0,0,1},
-  {1,0,0,0,0,0,2,0,0,0,2,0,0,0,0,0,0,0,0,0,0,0,0,1},
-  {1,0,0,0,0,0,2,2,0,2,2,0,0,0,0,3,0,3,0,3,0,0,0,1},
-  {1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1},
-  {1,0,0,0,0,0,2,2,2,2,2,0,0,0,0,0,0,0,0,0,0,0,0,1},
-  {1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1},
-  {1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1},
-  {1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1},
-  {1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1},
-  {1,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1},
-  {1,4,4,4,4,4,4,4,4,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1},
-  {1,4,0,4,0,0,0,0,4,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1},
-  {1,4,0,0,0,0,5,0,4,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1},
-  {1,4,0,4,0,0,0,0,4,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1},
-  {1,4,0,4,4,4,4,4,4,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1},
-  {1,4,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1},
-  {1,4,4,4,4,4,4,4,4,0,0,0,0,0,0,0,0,0,0,0,0,0,0,1},
-  {1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1,1}
-}; */
-
-/* void	draw_player(t_data *data, int x, int y)
-{
-	int hold;
-	int hold2;
-
-	hold = y + 9;
-	while (y < hold)
-	{
-		hold2 = x + 9;
-		while (x < hold2)
-			mlx_put_pixel(data->ray->img, x++, y, 0x8a15bdFF);
-		y++;
-		x -= 9;
-	}
-	y -= 9;	
-}
-
-void	draw_wall(t_data *data, int x, int y)
-{
-	int hold;
-	int hold2;
-
-	hold = y + 9;
-	while (y < hold)
-	{
-		hold2 = x + 9;
-		while (x < hold2)
-			mlx_put_pixel(data->ray->img, x++, y, 0xFFFFFFFF);
-		y++;
-		x -= 9;
-	}
-	y -= 9;
-}
-
-void	draw_space(t_data *data, int x, int y)
-{
-	int hold;
-	int hold2;
-
-	hold = y + 9;
-	while (y < hold)
-	{
-		hold2 = x + 9;
-		while (x < hold2)
-			mlx_put_pixel(data->ray->img, x++, y, 0x000000FF);
-		y++;
-		x -= 9;
-	}
-	y -= 9;
-}
-
-void	draw_map(t_data *data)
-{
-	int	i;
-	int	j;
-	int x;
-	int	y;
-
-	i = -1;
-	y = 0;
-	while (++i < mapHeight)
-	{
-		j = -1;
-		x = 0;
-		while (++j < mapWidth)
-		{
-			if (j == (int) data->ray->posX && i == (int) data->ray->posY)
-				draw_player(data, x, y);
-			else if (worldMap[i][j] != 0)
-				draw_wall(data, x, y);
-			else
-				draw_space(data, x, y);
-			x+= 10;
-		}
-		y += 10;
-	}
-} */
